Delete shard list on deactivate (so that next activate gets latest list)
After a Kinesis stream is resharded, “storm rebalance” command can be
invoked to refresh the shard list and distribute the latest shards
across the Spout tasks. “storm rebalance” first deactivates all the
tasks and then reactivates them. On deactivation, the shard list is
deleted from Zookeeper and on reactivation, the Spout fetches the shard
list because it is not present in Zookeeper.